Rename the plugins to Loaders and re-work the packages a bit.
